requirements.txt
altair==6.0.0
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
gitdb==4.0.12
GitPython==3.1.46
great_expectations==1.14.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
marshmallow==3.26.2
minio==7.2.20
mistune==3.2.0
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
ruamel.yaml==0.19.1
scipy==1.17.1
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
watchdog==6.0.0
